requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
beautifulsoup4==4.12.3
certifi==2025.1.31
charset-normalizer==3.4.0
contourpy==1.3.2
cycler==0.12.1
discord.py==2.5.2
distro==1.9.0
filelock==3.17.0
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2025.2.0
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.4.4
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
yarl==1.20.0
